+++ WorldScene.cpp
@@ -60,7 +60,7 @@ void CWorldScene::_ParseSection_ANIMATIONS(string line)
 	LPANIMATION ani = new CAnimation();
 
 	int ani_id = atoi(tokens[0].c_str());
-	for (int i = 1; i < tokens.size(); i += 2)	// why i+=2 ?  sprite_id | frame_time  
+	for (size_t i = 1; i < tokens.size(); i += 2)	// why i+=2 ?  sprite_id | frame_time  
 	{
 		int sprite_id = atoi(tokens[i].c_str());
 		int frame_time = atoi(tokens[i + 1].c_str());
@@ -175,9 +175,9 @@ void CWorldScene::_ParseSection_BACKGROUND_COLOR(string line)
 	// Skip invalid line
 	if (tokens.size() < 3) return;
 
-	float red = atof(tokens[0].c_str());
-	float blue = atof(tokens[1].c_str());
-	float green = atof(tokens[2].c_str());
+	float red = (float)atof(tokens[0].c_str());
+	float blue = (float)atof(tokens[1].c_str());
+	float green = (float)atof(tokens[2].c_str());
 
 	bg_color = D3DXCOLOR(red / 255, blue / 255, green / 255, 0.0f);
 }
@@ -263,7 +263,7 @@ void CWorldScene::Load()
 
 void CWorldScene::Unload()
 {
-	for (int i = 0; i < objects.size(); i++)
+	for (size_t i = 0; i < objects.size(); i++)
 		delete objects[i];
 
 	objects.clear();
@@ -278,11 +278,11 @@ void CWorldScene::Update(DWORD dt)
 void CWorldScene::Render()
 {
 	// Always render decorated objects first
-	for (int i = 0; i < decoratedObjects.size(); i++) {
+	for (size_t i = 0; i < decoratedObjects.size(); i++) {
 		decoratedObjects[i]->Render();
 	}
 
-	for (int i = 0; i < objects.size(); i++) {
+	for (size_t i = 0; i < objects.size(); i++) {
 		objects[i]->Render();
 	}
 }
